+++ jgclark.DashboardReact/src/moveClickHandlers.js
@@ -10,7 +10,7 @@ import { handlerResult } from './clickHandlers'
 import {
   getOpenItemParasForCurrentTimePeriod,
   getRelevantOverdueTasks,
-  getSettings,
+  getCombinedSettings,
   moveItemBetweenCalendarNotes,
 } from './dashboardHelpers'
 // import { showDashboardReact } from './reactMain'
@@ -45,7 +45,7 @@ const WEBVIEW_WINDOW_ID = `${pluginJson['plugin.id']}.main`
 // Instruction from a 'moveButton' to move task from calendar note to a different calendar note.
 export async function doMoveFromCalToCal(data: MessageDataObject): Promise<TBridgeClickHandlerResult> {
   const { filename, content, controlStr } = validateAndFlattenMessageObject(data)
-  const config = await getSettings()
+  const config = await getCombinedSettings()
   const dateInterval = String(controlStr)
   let startDateStr = ''
   let newDateStr = ''
@@ -102,7 +102,7 @@ export async function doMoveFromCalToCal(data: MessageDataObject): Promise<TBrid
 export async function scheduleAllYesterdayOpenToToday(_data: MessageDataObject): Promise<TBridgeClickHandlerResult> {
   try {
     let numberScheduled = 0
-    const config = await getSettings()
+    const config = await getCombinedSettings()
     // For these purposes override one config item:
     config.separateSectionForReferencedNotes = true
     const thisStartTime = new Date()
@@ -236,7 +236,7 @@ export async function scheduleAllTodayTomorrow(_data: MessageDataObject): Promis
   try {
 
     let numberScheduled = 0
-    const config = await getSettings()
+    const config = await getCombinedSettings()
     // For these purposes override one config item:
     config.separateSectionForReferencedNotes = true
     const thisStartTime = new Date()
@@ -364,7 +364,7 @@ export async function scheduleAllTodayTomorrow(_data: MessageDataObject): Promis
 export async function scheduleAllOverdueOpenToToday(_data: MessageDataObject): Promise<TBridgeClickHandlerResult> {
   try {
     let numberChanged = 0
-    const config = await getSettings()
+    const config = await getCombinedSettings()
     // For these purposes override one config item:
     config.separateSectionForReferencedNotes = true
     const thisStartTime = new Date()