+++ jgclark.DashboardReact/src/react/support/filterDropdownItems.js
@@ -1,24 +1,27 @@
 // @flow
 import { allSectionDetails, nonSectionSwitches } from "../../constants.js"
-import type {TDropdownItem,TSharedSettings} from "../../types.js"
+import type { TDropdownItem, TSharedSettings } from "../../types.js"
+import { getTagSectionDetails } from "./sectionHelpers.js"
 
-export const createDropdownItems = (
+export const createFilterDropdownItems = (
   sharedSettings: TSharedSettings,
-  tagSectionStr: string,
   pluginSettings: TAnyObject
 ): Array<TDropdownItem> => {
-  const dropdownSectionNames = allSectionDetails.filter(s => s.showSettingName !== '').map((s) => ({
-    label: `Show ${s.sectionCode !== 'TAG' ? s.sectionName : tagSectionStr}`,
+  const sectionsWithoutTag = allSectionDetails.filter(s => s.sectionCode !== 'TAG')
+  const tagSections = getTagSectionDetails(sharedSettings, pluginSettings)
+  const sectionsWithTags = [...sectionsWithoutTag, ...tagSections]
+  const dropdownSectionNames = sectionsWithTags.filter(s => s.showSettingName !== '').map((s) => ({
+    label: `Show ${s.sectionName}`,
     key: s.showSettingName,
     type: 'switch',
-    checked: (typeof sharedSettings !== undefined && sharedSettings[s.showSettingName]) ??  pluginSettings[s.showSettingName] ?? true,
+    checked: (typeof sharedSettings !== undefined && sharedSettings[s.showSettingName]) ?? pluginSettings[s.showSettingName] ?? true,
   }))
 
   const nonSectionItems = nonSectionSwitches.map(s => ({
     label: s.label,
     key: s.key,
     type: 'switch',
-    checked: (typeof sharedSettings !== undefined && sharedSettings[s.key]) ??  pluginSettings[s.key] ?? s.default,
+    checked: (typeof sharedSettings !== undefined && sharedSettings[s.key]) ?? pluginSettings[s.key] ?? s.default,
   }))
 
   return [...nonSectionItems, ...dropdownSectionNames]