+++ jgclark.DashboardReact/src/react/support/sectionHelpers.js
@@ -3,6 +3,18 @@ import { type TSection, type TSharedSettings, type TSectionCode, type TSectionDe
 import { allSectionDetails } from "../../constants.js"
 import { logDebug } from '@helpers/react/reactDev.js'
 
+const sectionWithTag = allSectionDetails.filter(s => s.sectionCode === 'TAG')[0]
+
+/**
+ * Get a consistent showSettingName for a given tag.
+ * @param {string} tag 
+ * @returns {string} The setting name.
+ */
+export function getShowTagSettingName(tag: string): string {
+  const showSetting = sectionWithTag.showSettingName
+  return `${showSetting}_${tag}`
+}
+
 /**
  * Gets the visibility setting for a given section code.
  * 
@@ -10,10 +22,11 @@ import { logDebug } from '@helpers/react/reactDev.js'
  * @param {TSharedSettings} sharedSettings - Shared settings to determine visibility of sections.
  * @returns {boolean} - Whether the section is visible.
  */
-const sectionIsVisible = (sectionCode: TSectionCode, sharedSettings: TSharedSettings): boolean => {
+const sectionIsVisible = (section:TSection, sharedSettings: TSharedSettings): boolean => {
+  const sectionCode: TSectionCode = section.sectionCode
   if (!sharedSettings) return false
-  const thisSection = getSectionDetailsFromSectionCode(sectionCode) // get sectionCode, sectionName, showSettingName
-  const settingName = thisSection?.showSettingName
+  // const thisSection = getSectionDetailsFromSectionCode(sectionCode) // get sectionCode, sectionName, showSettingName
+  const settingName = section.showSettingName
   if (!settingName) return true
   // TODO(later): alter this first part of the ternary to do something like startsWitth
   const showSetting = sectionCode === 'TAG' ? sharedSettings[settingName] : sharedSettings[settingName]
@@ -26,13 +39,14 @@ const sectionIsVisible = (sectionCode: TSectionCode, sharedSettings: TSharedSett
  * @param {TSection} section - The section to get the setting name for.
  * @returns {string} - The setting name.
  */
-export function getSettingName(section:TSection):string {
-  let settingName = getSectionDetailsFromSectionCode(section.sectionCode)?.showSettingName || ''
-  if (section.sectionCode === 'TAG') {
-    settingName = `${settingName}_${section.name}`
-  }
-  return settingName
-}
+// export function getSettingName(section:TSection):string {
+//   logDebug('sectionHelpers', `getSettingName ${section.sectionCode} ${section.name}`)
+//   const settingName = section.showSettingName ?? (getSectionDetailsFromSectionCode(section.sectionCode)?.showSettingName || '')
+//   // if (!settingName && section.sectionCode === 'TAG') {
+//   //   settingName = section.showSettingName
+//   // }
+//   return settingName
+// }
 
 /**
  * Reduce the useFirst array to include only the visible sections
@@ -40,13 +54,17 @@ export function getSettingName(section:TSection):string {
  *
  * @param {Array<TSectionCode>} useFirst - Priority order of sectionCode names to determine retention priority.
  * @param {TSharedSettings} sharedSettings - Shared settings to determine visibility of sections.
+ * @param {Array<TSection>} sections - The sections to filter.
  * @returns {Array<TSectionCode>} - Filtered and prioritized section codes.
  */
 function getUseFirstButVisible(
   useFirst: Array<TSectionCode>,
-  sharedSettings: TSharedSettings
+  sharedSettings: TSharedSettings,
+  sections: Array<TSection>
 ): Array<TSectionCode> {
-  return sharedSettings ? useFirst.filter((sectionCode) => sectionIsVisible(sectionCode, sharedSettings)) : useFirst
+  return sharedSettings ? 
+    useFirst.filter((sectionCode) => sections.find((section) => section.sectionCode === sectionCode) && sectionIsVisible(sectionCode, sharedSettings)) 
+  : useFirst
 }
 
 /**
@@ -67,7 +85,7 @@ export function getSectionsWithoutDuplicateLines(
   if (!paraMatcherFields) return _sections
   const sections = JSON.parse(JSON.stringify(_sections)) // Deep copy so we don't mutate the original pluginData
 
-  const useFirstVisibleOnly = getUseFirstButVisible(useFirst, sharedSettings)
+  const useFirstVisibleOnly = getUseFirstButVisible(useFirst, sharedSettings, sections)
 
   const orderedSections = useFirstVisibleOnly.map((st) =>
     sections.find((section) => section.sectionCode === st)
@@ -115,7 +133,7 @@ export const countTotalSectionItems = (sections: Array<TSection>): number => {
  */
 export const countTotalVisibleSectionItems = (sections: Array<TSection>, sharedSettings: TSharedSettings): number => {
   return sections.reduce((total, section) => {
-    if (sectionIsVisible(section.sectionCode, sharedSettings)) {
+    if (sectionIsVisible(section, sharedSettings)) {
       return total + section.sectionItems.length
     }
     return total
@@ -130,8 +148,22 @@ export const countTotalVisibleSectionItems = (sections: Array<TSection>, sharedS
  */
 export function getSectionDetailsFromSectionCode(thisSectionCode: string): TSectionDetails|void {
   const found = allSectionDetails.find((section) => section.sectionCode.startsWith(thisSectionCode))
-  if (!found) logDebug('sectionHelpers', `Section code: ${thisSectionCode} not found in allSectionDetails`)
+  if (!found) {
+    logDebug('sectionHelpers', `Section code: ${thisSectionCode} not found in allSectionDetails`)
+  }
   return found
+}
+
+/**
+ * Get Section Details for all tags in settings
+ * @param {TSharedSettings} sharedSettings 
+ * @param {TAnyObject} pluginSettings 
+ * @returns {Array<TSectionDetails>} {sectionCode, sectionName, showSettingName}
+ */
+export function getTagSectionDetails(sharedSettings: TSharedSettings, pluginSettings:TAnyObject): Array<TSectionDetails> {
+  //   { sectionCode: 'TAG', sectionName: '', showSettingName: `showTagSection` }
+  const tags = (sharedSettings.tagToShow ?? pluginSettings.tagToShow ?? '').split(',').map(t => t.trim()).filter(t => t !== '')
+  return tags.map(t => ({ sectionCode: "TAG", sectionName: t, showSettingName:getShowTagSettingName(t) }))
 }
 
 /**